cambio de nombre componentes
Modifique el nombre de algunos componentes para que haya cierta homogeneidad en relación con los otros
