+++ client/src/components/Vendedores/Dashboard/VendedoresDashboard.jsx
@@ -1,71 +1,350 @@
-import style from './CardPotencialLead.module.css';
+import style from "./VendedoresDashboard.module.css";
+import { useEffect } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { getAllLead } from "../../../redux/actions";
 import {
-	Card,
-	Table,
-	TableHead,
-	TableRow,
-	TableHeaderCell,
-	TableBody,
-	TableCell,
-	Text,
-} from '@tremor/react';
-import { CiCalendarDate } from 'react-icons/ci';
+  Card,
+  Table,
+  TableHead,
+  TableRow,
+  TableHeaderCell,
+  TableBody,
+  TableCell,
+  Text,
+  Title,
+  Badge,
+} from "@tremor/react";
 
-export const CardPotencialLead = () => {
-	const data = [
-		{
-			Name: 'Social Media Lab',
-			Category: 'Abogado',
-			Nivel: 1,
-			Thelephone: '+492563627',
-			Email: 'arroragaur@gmail.com',
-			Status: true,
-		},
-	];
+import Nav from "../../Nav/Nav";
 
-	return (
-		<Card className='h-screen bg-[39394b] w-[40rem]'>
-			<Table className={style.table}>
-				<TableHead className={style.tableHead}>
-					<TableRow className={style.tableRow}>
-						<TableHeaderCell className='text-start'>Name</TableHeaderCell>
-						<TableHeaderCell className='text-start ml-5'>
-							Profesion
-						</TableHeaderCell>
-						<TableHeaderCell className='text-start ml-5'>Nivel</TableHeaderCell>
-						<TableHeaderCell className='text-start ml-5'>Telefono</TableHeaderCell>
-						<TableHeaderCell className='text-start ml-5'>Email</TableHeaderCell>
-						<TableHeaderCell className='text-start ml-5'>Status</TableHeaderCell>
-					</TableRow>
-				</TableHead>
+const VendedoresDashboard = () => {
+  const { lead } = useSelector((state) => state);
+  const dispatch = useDispatch();
 
-				<TableBody className={style.tableBody}>
-					{data.map((item) => (
-						<TableRow
-							key={item.id}
-							className={style.tableCards}>
-							<TableCell className='flex justify-start items-center p-0'>
-								<Text className='text-start'>{item.Name}</Text>
-							</TableCell>
-							<TableCell className='flex justify-start items-center p-0'>
-								<Text className='text-start'>{item.Category}</Text>
-							</TableCell>
-							<TableCell className='flex justify-start items-center p-0'>
-								<Text className='text-start'>{item.Nivel}</Text>
-							</TableCell>
-							<TableCell className='flex justify-start items-center p-0'>
-								<Text className='text-start'>{item.Thelephone}</Text>
-							</TableCell>
-							<TableCell className='flex justify-start items-center p-0'>
-								<Text className='text-start'>{item.Email}</Text>
-							</TableCell>
-							<TableCell className='flex justify-start items-center p-0'>
-								<Text className='text-start'>{item.Status}</Text>
-							</TableCell>
-						</TableRow>
-					))}
-				</TableBody>
-			</Table>
-		</Card>
-	);
+  useEffect(() => {
+    dispatch(getAllLead());
+  }, [dispatch]);
+
+  return (
+    <div className="flex flex-row w-screen">
+      <Nav />
+      <Card className="w-full m-5 h-screen ">
+        <div className="flex justify-between items-center m-5">
+          <Title className={style.title}>Dashboard</Title>
+          <button className={style.buttonAdd}>Send </button>
+        </div>
+
+        <Table className={style.table}>
+          <TableHead className={style.tableHead}>
+            <TableRow className={style.tableRow}>
+              <TableHeaderCell className="text-start">
+                Invoice Id
+              </TableHeaderCell>
+              <TableHeaderCell className="text-start">Name</TableHeaderCell>
+              <TableHeaderCell className="text-start">
+                Profesion
+              </TableHeaderCell>
+              <TableHeaderCell className="text-start">Nivel</TableHeaderCell>
+              <TableHeaderCell className="text-start">
+                Telephone
+              </TableHeaderCell>
+              <TableHeaderCell className="text-start">Email</TableHeaderCell>
+              <TableHeaderCell className="text-start">Status</TableHeaderCell>
+            </TableRow>
+          </TableHead>
+
+          <TableBody className={style.tableBody}>
+            {lead.map((item, index) => (
+              <TableRow key={item._id} className={style.tableCards}>
+                <TableCell className="flex justify-start items-center p-0">
+                  <div className="w-8 ml-2 mr-4 rounded-full">{item._id}</div>
+                </TableCell>
+                <TableCell className="flex justify-start items-center p-0">
+                  <Text className="text-start">{item.name}</Text>
+                </TableCell>
+                <TableCell className="flex justify-start items-center p-0">
+                  <Text className="text-start">{item.profession}</Text>
+                </TableCell>
+                <TableCell className="flex justify-start items-center p-0 mx-3">
+                  <Text className="text-start bg-[#6254ff] py-2 px-3 text-xl rounded-md">
+                    {item.level}
+                  </Text>
+                </TableCell>
+                <TableCell className="flex justify-start items-center p-0">
+                  <Text className="text-start">{item.contact_number}</Text>
+                </TableCell>
+                <TableCell className="flex justify-start items-center p-0">
+                  <Text className="text-start">{item.email}</Text>
+                </TableCell>
+                <TableCell className="flex justify-start items-start p-0">
+                  <button className="bg-[#ff69b4] rounded-3xl px-6 py-1">
+                    Contratado
+                  </button>
+                </TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </Card>
+    </div>
+  );
 };
+
+export default VendedoresDashboard;
+
+// import { useState, useEffect } from 'react';
+// import style from './DashboardVendedores.module.css';
+
+// import {
+// 	Card,
+// 	Table,
+// 	TableHead,
+// 	TableRow,
+// 	TableHeaderCell,
+// 	TableBody,
+// 	TableCell,
+// 	Text,
+// 	Title,
+// 	Badge,
+// } from '@tremor/react';
+
+// import { CiGlobe, CiWarning, CiInstagram } from 'react-icons/ci';
+// import Nav from '../../Nav/Nav';
+
+// const DashboardVendedores = () => {
+// 	const [client, setClient] = useState([
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 	]);
+// 	const [nivel, setNivel] = useState({
+// 		1: '',
+// 		2: '',
+// 		3: '',
+// 		4: '',
+// 		5: '',
+// 		6: '',
+// 		7: '',
+// 		8: '',
+// 		9: '',
+// 		10: '',
+// 	});
+// 	const data = [
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 		{
+// 			InvoiceId: 2222222,
+// 			Name: 'Juan Carlos Gomez',
+// 			Profesion: 'www.google.com',
+// 			Nivel: 1,
+// 			Telephone: '+22131234',
+// 			Email: 'Hola@gmail.com',
+// 			Status: 'Contratado',
+// 		},
+// 	];
+// 	const handleNivelClick = (event) => {
+// 		setNivel({ ...nivel, [event.target.name]: event.target.value });
+// 	};
+// 	const handleSubmit = (event) => {
+// 		event.preventDefault();
+// 	};
+// 	return (
+// 		<div className='flex flex-row w-screen'>
+// 			<Nav />
+// 			<Card className='w-full m-5 h-screen '>
+// 				<div className='flex justify-between items-center m-5'>
+// 					<Title className={style.title}>Dashboard</Title>
+// 					<button className={style.buttonAdd}>Send </button>
+// 				</div>
+
+// 				<Table className={style.table}>
+// 					<TableHead className={style.tableHead}>
+// 						<TableRow className={style.tableRow}>
+// 							<TableHeaderCell className='text-start'>
+// 								Invoice Id
+// 							</TableHeaderCell>
+// 							<TableHeaderCell className='text-start'>Name</TableHeaderCell>
+// 							<TableHeaderCell className='text-start'>
+// 								Profesion
+// 							</TableHeaderCell>
+// 							<TableHeaderCell className='text-start'>Nivel</TableHeaderCell>
+// 							<TableHeaderCell className='text-start'>
+// 								Telephone
+// 							</TableHeaderCell>
+// 							<TableHeaderCell className='text-start'>Email</TableHeaderCell>
+// 							<TableHeaderCell className='text-start'>Status</TableHeaderCell>
+// 						</TableRow>
+// 					</TableHead>
+
+// 					<TableBody className={style.tableBody}>
+// 						{client.map((item, index) => (
+// 							<TableRow
+// 								key={item.InvoiceId}
+// 								className={style.tableCards}>
+// 								<TableCell className='flex justify-start items-center p-0'>
+// 									<div className='w-8 ml-2 mr-4 rounded-full'>
+// 										{item.InvoiceId}
+// 									</div>
+// 								</TableCell>
+// 								<TableCell className='flex justify-start items-center p-0'>
+// 									<Text className='text-start'>{item.Name}</Text>
+// 								</TableCell>
+// 								<TableCell className='flex justify-start items-center p-0'>
+// 									<Text className='text-start'>{item.Profesion}</Text>
+// 								</TableCell>
+// 								<TableCell className='flex justify-start items-center p-0 mx-3'>
+// 									<Text className='text-start bg-[#6254ff] py-2 px-3 text-xl rounded-md'>
+// 										{item.Nivel}
+// 									</Text>
+// 								</TableCell>
+// 								<TableCell className='flex justify-start items-center p-0'>
+// 									<Text className='text-start'>{item.Telephone}</Text>
+// 								</TableCell>
+// 								<TableCell className='flex justify-start items-center p-0'>
+// 									<Text className='text-start'>{item.Email}</Text>
+// 								</TableCell>
+// 								<TableCell className='flex justify-start items-start p-0'>
+// 									<button className='bg-[#ff69b4] rounded-3xl px-6 py-1'>
+// 										Contratado
+// 									</button>
+// 								</TableCell>
+// 							</TableRow>
+// 						))}
+// 					</TableBody>
+// 				</Table>
+// 			</Card>
+// 		</div>
+// 	);
+// };
+
+// export default DashboardVendedores;
